move cmd back to main and add comments
